requirements.txt
Flask==2.2.5
line-bot-sdk==2.2.1
google-cloud-secret-manager==2.10.0